requirements.txt
Flask
python-docx
PyMuPDF
PyPDF2
requests
beautifulsoup4
groq
python-dotenv
